aiviary-tree/app/nests/meta/enrichment-worker/requirements.txt
anthropic
google-cloud-aiplatform
google-cloud-speech
psycopg2-binary
pgvector
python-magic
requests
ffmpeg-python
